Save optional, arbitrary location identifier
Allow user to specify an arbitrary identifier for each location
by passing the name of a column in the input file. This identifier
is then saved along with any geolocation results.

This is useful for creating a table of locations that can be easily
joined with other database tables by, for example, passing the primary
key of another table at the `meta` identifier for the geolocation
results table.