Improve check for existing record
The former code considered any location already in the database to be
a duplicate. This update checks the location and the meta_id column
(if using the --meta flag to store an ID with locations) in order to
decide if a record already exists.

This allows two records with the same location, but different meta_id
values to have separate records in the database table.